Implemented a setWall feature to put a wall along a particular axis.
This has necessitated some changes to how PBC is handled. The cell now has a self.wall
attribute that either holds the cell dimension in that axis or 0 if there is a wall.

This is somewhat hacky as it relise on numpy returing nan's to indicate where the wall
is present, but it seems to work for the time being.